cli: add logging
Currently, when we receive user bug reports, we can't debug them, as we
don't get a detailed history of the internal state of our tool.

This commit adds logging to the CLI.

Logs are saved in the user's .bento directory. Right now we only store
the last tool run, because I didn't want to deal with log rotation yet.
In the future, we should probably store the last N runs.

Testing was manual. I ran the various bento CLI commands, ensuring that
log output was reasonable.