git: Add commit message template (#72)
This is a cherry-pick of commit 7348247d34:

This is an attempt to standardize our commit messages, in order to
guarantee a useful git history. The goal is for git blame to be
sufficient to understand why a line of code exists and what it does.

This commit adds a template file, .gitmessage, to the base of the repo.
In order to use it, a user must run:

  git config commit.template .gitmessage

From then on, commits will begin with this template.