js-tools: Install in yarn workspace root (#90)
yarn projects can be nested, with packages installed in sub-folders. By
default, calling `yarn add` in the workspace root will fail, so we need
to add a flag, `--ignore-workspace-root-check`, in order for yarn to
install eslint correctly.

I tested this by running `bento init` and `bento check` on the
`netlify-cms` project. We should add this as an integration test once
our integration test framework is available.