Add test for react project (#92)
* Add integration test for react project

We've had reports that bento does not work with jsx files. It turns out
that is too broad of a report, and it instead is that case that bento
does not work on an improperly installed checkout of
minimal-flask-react-app.

To validate that we work with .jsx files, I've added a test case to
validate that we do indeed correctly operate with react and .jsx.