metrics: Ensure always logged (#104)
* metrics: Ensure always logged

Previously, if a command threw an exception or called sys.exit(), we
would terminate immediately, and not log metrics. This is bad, as we
want to know when our users experience issues.

Here I use a functional decorator to wrap all subcommands. This
decorator handles all exceptions thrown by commands, including system
exits, guaranteeing proper metrics logging.

Note that the master "bento" command is _not_ decorated, as it would
cause a double logging given how click works. I plan to fix this in the
future.

I've added test cases to validate this.