tool_result: add caching to raw run output (#118)
Tools had to be rerun even if no files were changed since last run

Adds a simple cache for raw output of tools and checks cache before running tools

Added unit tests for run_cache

Instead of a single yaml file containing the metadata and data we have separated
them into two files. Also yaml is 100x slower to load than json so we use
json for the metadata instead. Also a bento version field was added
to the metadata so that we can invalidate the cache on version bumps.