run cache: Use file-mtime hash to determine change
This is a cherry-pick of 9c0719336.

Previously, we did not detect deleted directories as a file change. This
was because we never saw the mtime of those missing directories.

Here I use a murmur hash to calculate a hash of the directory, including
modification times.

This takes about 0.3 seconds on a large repo.

I updated test cases to include both no modifications and file removal.
Also updated to use tmp_path instead of tmpdir fixtures per pytest
recommendations.