bento: Ignore node_modules in python (#126) (#158)
This is a cherry-pick of 727f947b4a

node-gyp seems fairly common, and includes python files. If it is in
a node project, we will detect that project as python, and run checks on
it.

This commit excludes node_modules directories from Python tools.

This changes our acceptance tests quite a bit, so I've updated them.

I've also made the output expectations optional, as I discovered that we
need to install yarn prior to running tests.

Bonus: Exclude yarn packages from eslint.