Allow setting the base directory via --base-path. (#157)
This also tries to discover the base directory if Bento is run in a
subdirectory. This allows better CI and tooling integration.

As a side-effect, this lets users specify config options as environment
variables; --base-path is accessible as `BENTO_BASE_PATH`, and so on.

This is part of #145, though it doesn't totally fix it since currently there are
weird issues around cacheing and log output: we use absolute paths internally
now, so log messages are longer and caches break if the repository gets moved.