Add BENTO_EMAIL environment variable
In testing, we need a way to set the bento email without altering the
tester's user environment. We can not send --email ... to all bento
invocations, as there is no way to pass this flag into git commit.

Instead, add a BENTO_EMAIL environment variable, that can be used to
inject a test email address.