Post to /finding for findings (#240)
Lets use /metrics/finding endpoint so we log to a separate table. Everything flows to `_bento` index for Kibana which is the same as before! The event name changed from `bento_metrics_3` to `bento_finding_3`

 # Test
Run `bento check` and see data flowing to `bento_findings_3` for results data and  to `bento_metrics_3 ` for usage data in Segment Live debugger
<img width="1616" alt="Screen Shot 2019-11-07 at 4 47 42 PM" src="https://user-images.githubusercontent.com/1626382/68439958-62019780-017e-11ea-9731-01e4385fcff9.png">
<img width="1618" alt="Screen Shot 2019-11-07 at 4 47 38 PM" src="https://user-images.githubusercontent.com/1626382/68439959-629a2e00-017e-11ea-8057-16e42672315a.png">