flask: Add flake8-flask tool (#248)
This is our first integration of framework-specific checks. Here I add
a flask-specific tool (which itself is just an extension of our existing
flake8 tool).

In its current incarnation, this tool matches no projects, so it is
disabled by default. It can be enabled by adding 'r2c.flask' as a tool
to the user's '.bento.yml'.