ignores: Include gitignore (#283)
For the vast majority of users, any pattern in the .gitignore file
can be ignored by bento. Configure bento to include .gitignore patterns
by default.

To do this, I add a ":include path/to/other/file" directive that can be
used within .bentoignore.

To make all this easier to understand, I've split the code up to add
separate Parser and Processor classes.

In order to properly include gitignore patterns prior to project
detection, I flip the order of config and ignore loading when running
the CLI. This has the added bonus of removing a spurious "Using default
ignore patterns" messages when running on an uninitialized project.

I also add a fair amount of documentation to the default .bentoignore
file.

Test cases are added for all additional functionality.

In order to get the create-react-app test to pass appropriately, I have
to tell it to ignore 'packages/'.

* Fix stylish for CI

CI shows up as a zero-width terminal and a tty (which makes no sense,
since it's not a tty, but whatever). For this case, don't trim to the
terminal width.