tools: Add typed returns (#303)
Some tools benefit from an intermediate data representation that is not
a string. E.g., prior to this commit, we were double-parsing eslint
return data. We can also imagine use cases where a single string causes
memory issues within the CLI.

Here I allow non-string data representations by adding results-type
parameterization to both tools and parsers.

Note that parsers can be contravariant with respect to the results type,
as they are pure consumers of the results. Tools must be invariant as
they are both producers and consumers (via the cache).