check: Add --show-archived (#329)
We will be auto-archiving on init, so we need to give users a way to
view the archived issues (as they won't have a chance to view them prior
to constructing the archive).

To accomplish this, I add a --show-archived flag to bento check. I also
direct users to run this when we surface archived findings.

Test cases are updated to match this change.