make bento work with space in path (#343)
bugfix: allow bento to init and run over folders with paths in the name

returntocorp/bento#183

Wrap the commands with either single or double quotes to handle the space in bash command. Also adding a hack for flake8 to use '*' for spaces because of a bug

Manually tested it by running `bento init` and `bento check` over a folder with a space.