Add ability to run with multiple formatters; fix clippy (#353)
This commit allows a user to run with multiple formatters at once. This
is helpful, e.g., for displaying both histogram and detail output.

Here I update the default behavior to print both clippy and histogram
output.

As a bonus, I also update "next step" tips for consistency and
usability.

Finally, I also prevent paging if there are not at least a minimum
number of pages (in this case, 3), to prevent paging too early.