Content inventory (#423)
To first order, this is fundamentally a refactor commit. It moves
content and styling into separate content yaml files, while logic
remains within the CLI command code.

Every attempt was made to maintain the same output as already existed
for the registration and init workflows. However, these are not fully
covered by test code, so some of this validation was manual.

Also fixes a link rendering issue, where links were improperly formatted
when the link anchor text was split by a newline.