metrics: Separately anonymize results + usage data (#454)
Our usage data contain clearly identifying information (the user
email). However, they also contain data that are copied in the results
data (the repository and commit hashes). This means that we are breaking
anonymization.

To solve this, use separate hash salts for usage and results data.

Salts were generating using random keygen.

I also remove the redundant "repository" field. Note that we keep the
"hash_of_" name, as it makes it clear in our privacy documentation that
we are indeed hashing the data.

Also remove some unnecessary parends, and unused exception argument.