Move content inventory over to Python classes (#448)
Currently, our content inventories (for text emitted to the screen) are
written in YML files. As @ash pointed out, this means we can not use
type-checking to ensure content is well-formed, and we can not
deduplicate oft-repeated content. Furthermore, documentation can be
annoying, as it's not obvious where to look for the YML schema.

Here I use Python classes instead to construct the inventory.