qa tests: Don't print upgrade notice (#465)
Our QA tests currently fail if run on an older release branch, as the
upgrade output does not match expected output. This
creates issues if we ever want to deploy patch versions.

Here I add a BENTO_TEST environment variable. If set to some bash-truthy
value, it suppresses upgrade checking.

As a bonus:
- Unify private-method naming in bento/cli.py
- Deduplicate code in tests/acceptance/qa.py