init: Install non-blocking hook (#507)
We want to encourage users to use Bento on every commit. To do this, we
add functionality to install (and remove) Bento as a non-blocking commit
hook.

In addition, Bento is configured as a non-blocking commit hook, by
default, on init.

Note that git commit hashes are sensitive to the commit time, so
I removed them in our acceptance test output.