Send Exception class in metrics when Bento exits with an error (#515)
* Sends exception class in metrics when Bento exits with an error

Fixes the issue of not knowing what exception occurred or if the user interrupted the process manually.

KeyboardInterrupt is caught separately. Logging for keyboard interrupts is updated. The name of the exception class is sent in metrics.

Send the name of the exception class along with metrics when Bento errors out.

* address feedback

* Address feedback

* exception_name instead of exception in metrics