init: Rewrite content; install tools (#529)
This commit rewrites the content displayed to users on init,
per language from @luke and @dannibal.

The main point of this work is to move the user to a "commit first"
workflow, rather than causing the user to focus on technical debt.

We want to ensure commit is reasonably fast even on first use of Bento.
To do this, pre-install tools at init time.

To make this work, I've heavily refactored the ToolRunner class,
deduplicating code associated with progress bars.

In addition, I removed the setup latch, which was previously necessary
to prevent issues with the cache crawling venv directories. Since these
directories are now located outside the project, and since the cache
always ignores the .bento directory, this is unnecessary.