check: Don't wipe untracked files (#574)
If a change deletes a tracked file, and the file is then added back on
the filesystem, it will be untracked. This means the file is not stashed
by the pre-commit library, and it will be deleted when bento check is
run.

Here I simply hard abort bento check in this case, and then print next
steps for the user.

I've tested this manually. It should be added to our acceptance tests.

This commit addresses issue #571.