target_file_manager: refactor out file logic (#634)
* target_file_manager: refactor out file logic

Logic that determined what files to run on were very tightly coupled
with bento tool running and even the click context used in bento.

This PR tries to pull out as much of the file logic as it can so that
after the first few lines of archive/check we can safely assume that we are
passed all the paths to all the files we need to worry about. To this end
we create a TargetFileManager class that handles all the logic related to

a) Parsing .bentoignore
b) Traversing the project repo and listing all files
c) Ignoring any directory/file in .bentoignore patterns
d) Filtering the relevant files to paths specified in bento commands
e) Checking which files are staged
f) Doing any relative path resolution in case bento is called in a subdirectory

With this refactor it was also possible to limit how much path traversal bento
was doing to just paths relevant to the check/archive command. This means that
in steady state, bento check in a large project should no longer need to touch
all files.