run_cache: use cache when running with --all (#643)
Previous PRs had a regression on not using cache when running with --all. This
has bento using cache on --all again.

Note that even if we set use_cache to true for staged runs because
git resets the last touched timestamp of files when they are touched by
git our staged runs do not benefit from any caching.