tool: Don't exceed max arg limit (#642)
In 0.8, we explicitly pass the paths to check as an argument list. If
the number of files to check exceed the OS's built in max. arg list, the
tool will crash.

Here, we iterate over these files instead.

This commit is only manually tested, although it works on the zulip
repo, which requires multiple batches.