Fix commit-message validation logic (#738)
Prior to this commit, if the entire PR consisted of subject-only
commits, we would detect the PR as having no commits at all.

This commit also looks at subject lines when evaluating if a PR adds
commits. Note that I did not want to use git referential identity, as
this should not fire on a PR that (uselessly) attempts to merge an item
from the master branch's history.

Changes were validated by running on a PR with three subject-only
messages.