context: Error on bento's exit if an enabled tool isn't found (#743)
* util: Allow echo functions to print newlines

Prior to this commit, if you tried to echo multiple lines as an error or
such, the newlines would just get trimmed, breaking text formatting. An
example of such an error is the NoToolsConfiguredException, which
prompts the user to contact support on a separate, second line.

This commit changes our usage of textwrap to not do that anymore, and
instead, wrap text line-by-line. To prevent too much code duplication,
the commit also refactors echo_error, echo_warning, and echo_success to
share their code.

Changes were validated by calling echo_error with a multi-line string.
Test cases can easily be added, in case the change is considered useful.

* context: Allow deferring errors till exit

Prior to this commit, if you wanted to fail with an error but still
print useful results first, such as one tool not being found, you had to
resort to using an `is_error` flag or similar hacks.

This commit allows you to queue exceptions on the context which will be
raised once your invoked command finishes.

Changes were validated by adding two errors to the queue and checking
whether they're printed on exit.

* context: Error on bento's exit if an enabled tool isn't found

Prior to this commit, one could have a tool enabled in the project
configuration and Bento would just print a note that it couldn't be
found. Closes https://github.com/returntocorp/bento/issues/281

This commit changes it so that Bento has a failing exit code in these
cases, though it doesn't block the execution of other tools.

Changes were validated by adding two non-existent checks to the config
and running a `bento check`.