Deploy to PyPI on git tag
Prior to this commit, deployment to PyPI was a manual process.

This commit runs 'make release' when a git tag matching a specific
format is pushed.

Changes were validated by testing this next time we have to release. We
can also attempt a beta release in the near future if we'd like to
validate quicker.