Move sgrep docker logic to shared DockerTool class. (#834)
Prior to this commit, docker usage was reimplemented in each tool. For
ease of maintainability, let's solve Docker usage in one place only.

This commit moves Docker code to a new DockerTool class.

It also allows the use of BENTO_REMOTE_DOCKER as an alias for
R2C_USE_REMOTE_DOCKER.

Changes were validated by ensuring tests pass, as well as manually
running sgrep with various registries and values of
BENTO_REMOTE_DOCKER.

As a bonus, this reduces the amount of time it takes to run shellcheck
on all of echelon-backend from 33 seconds to 6 seconds.

Bonus: Fix issue with staged addition in project subdirectory. When a file
was added to the git index of a Bento project nested in a git subdirectory,
Bento would crash.