Add Gosec tool (#836)
This commit adds a gosec tool (disabled by default).

Unfortunately gosec can not be passed a file list. This means that the
runtime of this tool is constant in the length of the file list, and
linear in the project size, regardless of arguments (unless no Golang
files are selected).

Changes were validated by adding additional test cases.